Add strip test, small tests refactor
