+++ tests/TestCase.php
@@ -4,6 +4,16 @@ use Orchestra\Testbench\TestCase as TestBenchTestCase;
 
 class TestCase extends TestBenchTestCase
 {
+    /**
+     * Setup the test environment.
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->shortcode = app()->make('shortcode');
+    }
+
     protected function getPackageProviders($app)
     {
         return ['Webwizo\Shortcodes\ShortcodesServiceProvider'];
